Implemented Authentication

See merge request juzodinma/levelup-2023-autumn!7

+++ backend/src/main/java/dev/levelupschool/backend/service/CommentService.java
@@ -1,15 +1,22 @@
 package dev.levelupschool.backend.service;
 
+import dev.levelupschool.backend.auth.AuthenticationProvider;
+import dev.levelupschool.backend.dtos.CommentDto;
 import dev.levelupschool.backend.exception.ModelNotFoundException;
+import dev.levelupschool.backend.model.Article;
 import dev.levelupschool.backend.model.Comment;
+import dev.levelupschool.backend.model.User;
 import dev.levelupschool.backend.repository.ArticleRepository;
 import dev.levelupschool.backend.repository.CommentRepository;
-import dev.levelupschool.backend.repository.UserRepository;
 import dev.levelupschool.backend.request.CreateCommentRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PagedResourcesAssembler;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.hateoas.PagedModel;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import java.nio.file.AccessDeniedException;
 
 @Service
 public class CommentService {
@@ -18,46 +25,61 @@ public class CommentService {
     private CommentRepository commentRepository;
 
     @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
     private ArticleRepository articleRepository;
 
-    public List<Comment> getAllComments() {
-        return commentRepository.findAll();
+    @Autowired
+    private AuthenticationProvider authProvider;
+
+    @Autowired
+    private PagedResourcesAssembler<CommentDto> pagedResourcesAssembler;
+
+    public PagedModel<EntityModel<CommentDto>> getAllComments(Pageable paging) {
+        var result = commentRepository.findAll(paging).map(CommentDto::new);
+        return pagedResourcesAssembler.toModel(result);
     }
 
-    public Comment getComment(Long id) {
-        return commentRepository
+    public CommentDto getComment(Long id) {
+        return new CommentDto(commentRepository
             .findById(id)
-            .orElseThrow(() -> new ModelNotFoundException(id));
+            .orElseThrow(() -> new ModelNotFoundException(Comment.class, id)));
     }
 
-    public Comment createComment(CreateCommentRequest request) {
-        Long userId = request.getUserId();
-        Long articleId = request.getArticleId();
-        String content = request.getContent();
+    public CommentDto createComment(CreateCommentRequest request) {
+        User loggedInUser = authProvider.getAuthenticatedUser();
 
-        return commentRepository.save(
-            new Comment(
-                content,
-                userRepository.findById(userId).orElseThrow(() -> new ModelNotFoundException(userId)),
-                articleRepository.findById(articleId).orElseThrow(() -> new ModelNotFoundException(articleId))
-            )
-        );
+        Article article = articleRepository.findById(request.getArticleId())
+            .orElseThrow(() -> new ModelNotFoundException(Article.class, request.getArticleId()));
+
+        Comment comment = new Comment(request.getContent(), loggedInUser, article);
+
+        return new CommentDto(commentRepository.save(comment));
     }
 
-    public Comment updateComment(Comment newComment, Long id) {
-        return commentRepository
-            .findById(id)
-            .map(comment -> {
-                comment.setContent(newComment.getContent());
-                return commentRepository.save(comment);
-            })
-            .orElseThrow(() -> new ModelNotFoundException(id));
+    public CommentDto updateComment(Comment newComment, Long id) throws Exception {
+        var loggedInUser = authProvider.getAuthenticatedUser();
+
+        var comment = commentRepository.findById(id)
+            .orElseThrow(() -> new ModelNotFoundException(Comment.class, id));
+
+        if (!comment.getAuthor().getId().equals(loggedInUser.getId())) {
+            throw new AccessDeniedException("You cannot edit a comment not created by you!");
+        }
+
+        comment.setContent(newComment.getContent());
+
+        return new CommentDto(commentRepository.save(comment));
     }
 
-    public void deleteComment(Long id) {
+    public void deleteComment(Long id) throws Exception {
+        var loggedInUser = authProvider.getAuthenticatedUser();
+
+        var comment = commentRepository.findById(id)
+                .orElseThrow(() -> new ModelNotFoundException(Comment.class, id));
+
+        if (!loggedInUser.getId().equals(comment.getAuthor().getId())) {
+            throw new AccessDeniedException("You cannot delete an article not created by you!");
+        }
+
         commentRepository.deleteById(id);
     }
 }
